test: localize table fixtures to specific test suite
